requirements.txt

gunicorn
line-bot-sdk
pygsheets
Flask==2.2.2
